Revert "move apiQuery to a common place"
This reverts commit 1c53a70d2c1fff5468bba20794de4dd9a45dbc2b.
